Rename CurrentTrip component and partial file
