+++ src/Components/CurrentTrip/_currenttrip.scss
