+++ src/index.scss
@@ -7,6 +7,6 @@
 @import 'Components/MainView/mainview';
 @import 'Components/Form/form';
 @import 'Components/TripETA/trip-eta';
-@import 'Components/TripView/tripview';
+@import 'Components/CurrentTrip/currenttrip';
 @import 'Components/AddContact/_addcontact';
 @import 'Components/MapDisplay/map';
